:edit: updated projects to match submission projects.

+++ cc1-printf/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gonca <gonca@student.42.fr>                +#+  +:+       +#+        */
+/*   By: goperez- <goperez-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2026/03/29 17:35:09 by gonca             #+#    #+#             */
-/*   Updated: 2026/03/29 18:27:34 by gonca            ###   ########.fr       */
+/*   Created: 2026/03/29 17:35:09 by goperez-          #+#    #+#             */
+/*   Updated: 2026/05/14 16:03:21 by goperez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,9 +22,8 @@ int	ft_printf(const char *format, ...);
 // Utility functions
 int	ft_print_char(int c);
 int	ft_print_str(char *str);
-int	ft_print_digit(long n, int base, int uppercase);
+int	ft_print_digit(int n, int base, int uppercase);
 int	ft_print_ptr(unsigned long long ptr);
 int	ft_print_unsigned(unsigned long long n, int base, int uppercase);
-
 
 #endif
